+++ utils/banner.js
@@ -5,7 +5,14 @@ const Package = require('../package.json');
 
 // Banner function
 function banner() {
-  return `Lime version ${Package.version}\n${Package.description}\n2021 ${Package.author}, ${Package.license} license\n${Package.homepage.slice(0, -7)}`;
+  const lines = [
+    `Lime version ${Package.version}`,
+    Package.description,
+    `2021 ${Package.author}, ${Package.license} license`,
+    Package.homepage.slice(0, -7),
+  ];
+
+  return lines.join('\n');
 }
 
 /* ------------------------ division ------------------------ */